+++ app/src/main/java/com/example/cookpad/fragments/FragmentThree.kt
@@ -1,15 +1,18 @@
-package com.example.cookpad
+
+package com.example.cookpad.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.TextView
+import com.example.cookpad.R
 
 
-class FragmentTwo : Fragment() {
+class FragmentThree : Fragment() {
 
-
+    lateinit var tv_fragment_name: TextView
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -19,7 +22,10 @@ class FragmentTwo : Fragment() {
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        tv_fragment_name.text="Fragment Two"
+
+        tv_fragment_name = view.findViewById(R.id.tv_fragment_name)
+
+        tv_fragment_name.text="Fragment Three"
     }
 
 }
