+++ app/src/main/java/com/example/cookpad/fragments/FragmentTwo.kt
@@ -1,4 +1,4 @@
-package com.example.cookpad
+package com.example.cookpad.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,12 +6,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import org.w3c.dom.Text
+import com.example.cookpad.R
 
 
-class FragmentFour : Fragment() {
+class FragmentTwo : Fragment() {
 
-
+    lateinit var tv_fragment_name: TextView
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -21,10 +21,8 @@ class FragmentFour : Fragment() {
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-
-        tv_fragment_name.text="Fragment Four"
+        tv_fragment_name = view.findViewById(R.id.tv_fragment_name)
+        tv_fragment_name.text="Fragment Two"
     }
-
-
 
 }
