+++ app/src/main/res/layout/fragment_four.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FragmentFour">
+    tools:context=".fragments.FragmentFour">
 
     <TextView
         android:id="@+id/tv_fragment_name"
