+++ app/src/main/res/layout/fragment_two.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FragmentTwo">
+    tools:context=".fragments.FragmentTwo">
 
     <TextView
         android:id="@+id/tv_fragment_name"
